refactor(utils): moving util functions to utils directory
