+++ src/utils/isSymbol.ts
