+++ src/utils/memoize.ts
@@ -1,4 +1,4 @@
-import { MemoizeInterface } from './types';
+import { MemoizeInterface } from '../types';
 
 /**
  * Creates a function that memoizes the result of `func`. If `resolver` is
